ES6/ES2015/Babel + Webpack compatibility
The changes inside this library have been made in
order to solve a transpiling issue with Babel.

We are using this lib in combination of Babel and
Webpack, and we had an issue during the transpiling
process. Babel sadly does not understand the meaning
of `this`, so it was replaced with the next best it
could find, which was `undefined`. In order to
maintain compatibility with browser as well as other
environments, we decided to wrap the entire scope
inside a simple factory.